feat(essentia): 학사 다이제스트 발송을 디스코드로 전환
텔레그램 봇을 전부 폐기해 릴레이가 발송 불가 상태였다. 봇 토큰 대신
디스코드 채널 웹훅을 쓴다. 웹훅은 지정 채널 쓰기 권한만 가져 유출 시
피해가 그 채널로 한정되고, 폐기도 웹훅 삭제 한 번으로 끝난다.

telegram-relay 를 discord-relay 로 개명한다. 디렉토리·이미지·시크릿·
Service·빌드 워크플로 전부 포함.

포맷을 텔레그램 HTML 에서 디스코드 마크다운으로 옮긴다.

  <b>x</b>            → **x**
  <i>x</i>            → *x*
  <code>x</code>      → `x`
  <a href="u">↗</a>   → [↗](<u>)

링크를 <> 로 감싸 임베드 미리보기를 막는다 (텔레그램의
disable_web_page_preview 와 같은 효과). 과목명에 밑줄이 흔해
(GEDB002_44 등) 이스케이프 대상에 마크다운 특수문자를 포함했다.

디스코드는 메시지가 2000자로 제한된다 (텔레그램은 4096자). 과제가 많은
날 잘리지 않도록 줄 경계로 쪼개 순차 발송한다.

검증: 로컬에서 릴레이를 띄우고 실제 과목명이 든 샘플 다이제스트를
POST 해 디스코드 채널 도착까지 확인. kustomize build 및
kubectl apply --dry-run=server 통과.
