+++ .github/workflows/essentia-discord-relay-build.yaml
@@ -1,12 +1,12 @@
-name: essentia-telegram-relay build
+name: essentia-discord-relay build
 
 on:
   push:
     branches: [main]
     paths:
-      - k8s/essentia/telegram-relay/Dockerfile
-      - k8s/essentia/telegram-relay/relay.mjs
-      - .github/workflows/essentia-telegram-relay-build.yaml
+      - k8s/essentia/discord-relay/Dockerfile
+      - k8s/essentia/discord-relay/relay.mjs
+      - .github/workflows/essentia-discord-relay-build.yaml
   workflow_dispatch:
 
 permissions:
@@ -30,14 +30,14 @@ jobs:
       - uses: docker/metadata-action@v5
         id: meta
         with:
-          images: ghcr.io/manamana32321/essentia-telegram-relay
+          images: ghcr.io/manamana32321/essentia-discord-relay
           tags: |
             type=raw,value=latest
             type=sha,prefix=sha-,format=short
 
       - uses: docker/build-push-action@v5
         with:
-          context: k8s/essentia/telegram-relay
+          context: k8s/essentia/discord-relay
           push: true
           tags: ${{ steps.meta.outputs.tags }}
           labels: ${{ steps.meta.outputs.labels }}
