+++ k8s/essentia/discord-relay/sealed-secret.yaml
@@ -1,0 +1,12 @@
+apiVersion: bitnami.com/v1alpha1
+kind: SealedSecret
+metadata:
+  name: essentia-discord-relay
+  namespace: essentia
+spec:
+  encryptedData:
+    DISCORD_WEBHOOK_URL: AgBmiiSGREfR8aAuLl0D19PjOYXO/BnI2Xa8kE4ItYi9121V1vj0iRl0CGuDBdZuP8XP8yASH1m3/iZlRsalTfjPfyk1NWsY6tSyLy6yBuZiLcG0Ce/PEoRqTUBqG3mBj5ZkoFMVcgGxb0c9abmeMgYm2CuPqZtCqeEv0YO8G1BOiMCv3I6dQG329SR6GHSh/4+O/lHjoxibbxTn6aX9ECJYvDVgJ47NHwdydSrXq7bs4xkaGO4NuCLW5eqBhKYjt/xEe3nHiUFUSlywLBzmS1u2LE6A2nY39ql0OD2dQYUqfCiANqA9xmle5+xXuluPnT0AMmyw2RozZb23fcxg5Bxh0TvMlrVBePGQBbWWDABkp6Z2vnezQTMRwcNfY/08AIIz6Xg40rpydoJtxF9QucMgj6FuJopB7cRTg9CD1IgxbtXrEE8xkmEBOa/lsloveQHN2gYpa0C9Rcy0XKGxKOA62Q9vFsLXPuiwuMFN7P3HfrFbjC116pVudJfrG5tguQXuSn4Lvw5bxh3lZXZu/p0gX6PI9b/zAtvAsnSFHaBaskj2E+uuKjwRAi6NOma97gS+o18EI0aZI7s9XlGGylIJ0B80u3JRtTQJWdURb/prVRvgrlk/CyMkDhEKPRJnyR9zxE09wHHG33AYZs5QjswvZst86yRjpnRUh9ZPZ+Nb95F+LWGgEigfsnVqNZEAe5sA4tHa+QwLb5YKeK2R2M/8fKuDs3oVJiVnTs6hEWuyGORWwloiLqYOnQDK3yTByUEHRlYpL48ggV5mc+5rrxNpAZd7R+FHtqtcvbb22RgfAuQMrJOdj63k38MclJS2991dQw0vhE1vFuaNJiOmAzSBcrEw0urJZvFm
+  template:
+    metadata:
+      name: essentia-discord-relay
+      namespace: essentia
