+++ k8s/essentia/kustomization.yaml
@@ -3,13 +3,13 @@ kind: Kustomization
 # 2026-07-22 전면 정지. 사용 빈도가 낮은데 GHCR PAT 만료(2026-07-18)로 Image
 # Updater 만 4일째 에러를 뿜고 있었다. 데이터는 PVC(essentia-postgres-data)와
 # MinIO Tenant 에 그대로 남는다 — 되살릴 때 아래 replicas 블록만 지우면 된다.
-# 주의: 매일 07:00 KST 텔레그램 다이제스트도 함께 멈춘다 (telegram-relay 앱 내부 스케줄러).
+# 주의: 매일 07:00 KST 디스코드 다이제스트도 함께 멈춘다 (essentia-api 가 발송을 건다).
 replicas:
   - name: essentia-api
     count: 0
   - name: essentia-web
     count: 0
-  - name: essentia-telegram-relay
+  - name: essentia-discord-relay
     count: 0
   - name: essentia-postgres
     count: 0
@@ -26,9 +26,9 @@ resources:
   - web/deployment.yaml
   - web/ingress.yaml
   - web/service.yaml
-  - telegram-relay/deployment.yaml
-  - telegram-relay/sealedsecret.yaml
-  - telegram-relay/service.yaml
+  - discord-relay/deployment.yaml
+  - discord-relay/sealed-secret.yaml
+  - discord-relay/service.yaml
   - ghcr-pull-sealed-secret.yaml
   - canvas-credentials.yaml
 
@@ -39,5 +39,5 @@ resources:
 images:
   - name: ghcr.io/essentia-edu/api
     digest: sha256:1c30ed47e532f84b3a791e23a782baa45878b79a9d32d18b6712a9d39518b67a
-  - name: ghcr.io/manamana32321/essentia-telegram-relay
+  - name: ghcr.io/manamana32321/essentia-discord-relay
     newTag: latest
