--- k8s/essentia/telegram-relay/sealedsecret.yaml
@@ -1,14 +1,0 @@
-apiVersion: bitnami.com/v1alpha1
-kind: SealedSecret
-metadata:
-  name: essentia-telegram-relay
-  namespace: essentia
-spec:
-  encryptedData:
-    TELEGRAM_BOT_TOKEN: AgCO7EqFsJEy3j8r9/GbjmqbBIFzbEU/Zco5rueqNmiBI2x7R+TPdYFB7hbtNM0dvb4KLnCV3Gd5Tvxe88rcEKhyJL3zrg9FBvaJfXgmmrEz3rly1m85t2RA/B7Pgo2OIu1IdLDcmZ7Xs7SAN8SdRKeB4u3cekt0Y37orYZPoU5xzG+VhIkUXD0Gm+Xgx7KvnNggxn3NMq43jhPUTuHWLk6ivkOBz4wXP5weGm9oWZMwEQ98sCAGR0lBJDdW7y1/DPHUzIV90dRUt/UTiIUg1UsWJ7vVxuqWxc0KLAarPmYv4BmVAQQ48X7PlYNLIByXGHQ8O7rSh08Ddh2DoNNNYidbkPpLF9ty9i8KFIObb8tg6iJkYrkEshX3tSFPNZv2xN7D53i23NEjFPzlV7GGX6ZU5Ts57A/VzMe7RCzSUDasnzwQBdE8L7Gwn1C1MH9OdGaefyMwx+1Y0Vu2+f6PqjwI5YCAjFA5uGU3P9s/zWvrA3vL2a2Ls/062llnfPMyi8L1V4s6YZKffSQTr2cEuRC7XmqTFtee2lWFx2ozqqgI6WTuAkCS++aJVI4dTGj4fvkqFGit1GqeCrQRZfPeiqbyRu8S7MBt2EVV+LmpEqMJLvk140bU+g4Qbjr8vKOuuaB+3JnebpZMA+AGyF7XfsXsyRJOG2+OamNRX/VGJl3wvrp4Y9AklHtTxJ2SgQVe8f5XXyQFqdvTw4W7PA44fwofZ7LVXIEp+g/KwwAevaYdYdqEqADXlHZJ5jcRIFBf
-    TELEGRAM_CHAT_ID: AgB41VP9W3wlsn6SYMGPmVAYE25M4wFEluUOpij8iFoSPnivFasOxNstOoJJ3tx1M9JE9epZqnh3YFAE74Y0epKWjZA6LLgwpn4PPi6IHayLJH0QiXaL17a5MZXOwlXxQS8SMn/wetK0ab7oTU4PRgtfWNkWB8M6F9Gl+oUNXfHMLUGNZJf/SecirtTuerD5iuGIVWLZcWmVqt5WtpuosL5rq9qZGW9H/KIAo3+BEJWojKPSc1zoLKcXElS1cpwbp0569k9fYonbmsx9qFYO4lGwsx263lI5ccbfnIdyTUGqKl3fHCD/YXPx0JXcNtXbwMxnA1w8qCizQQ76SSOc5hu4gTMCkWNnGUiHm0yWPgRBW3kVl8J5f+zEj0Ofx//HawDy+kx1cXgE4IdTuNZnEq0h4zfnNyJM6CVTzqZZqG+am674lIDWzGP/W/LMI04JgG82mEJPKIGjlqnRb2J/dvNtnfkPElbBr/GzKAUAseTB4p4APZOwrLNgTfhJwBMvfFF1sNIEVcQV83KuBCYmhWu81mgMNnONUxlzRb4nJyYXe4IhqbjAlfXIzI9LYIuFwGWn1nzn7yB/cxRMOFvy7GbeqoOoloPUZyHMQGqPvP4FSN/ingNFNiyZwVmvDrszaWdfeWVQ8i4uBZmOuxxAbicCEcMD+qBE/pVgUBtBhGfW+B4slQSUUEj4N1RhT7GWs2Snkz/mnhp5HwNz
-  template:
-    metadata:
-      name: essentia-telegram-relay
-      namespace: essentia
-    type: Opaque
